Changing packages, now using a Vertical approach instead of 'bags'
