+++ Project/app/src/main/java/edu/isel/ion/android/courses/CoursesFragment.kt
@@ -1,4 +1,4 @@
-package edu.isel.ion.android.fragments.course_details
+package edu.isel.ion.android.courses
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -10,14 +10,14 @@ import edu.isel.ion.android.R
 /**
  * A simple [Fragment] subclass.
  */
-class CourseDetailsFragment : Fragment() {
+class CoursesFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_course_details, container, false)
+        return inflater.inflate(R.layout.fragment_courses, container, false)
     }
 
 }
