+++ Project/app/src/main/java/edu/isel/ion/android/home/HomeFragment.kt
@@ -1,13 +1,10 @@
-package edu.isel.ion.android.fragments.home
+package edu.isel.ion.android.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProviders
 import edu.isel.ion.android.R
 
 class HomeFragment : Fragment() {
