+++ Project/app/src/main/java/edu/isel/ion/android/schedule/ScheduleFragment.kt
@@ -1,6 +1,7 @@
-package edu.isel.ion.android.fragments.courses
+package edu.isel.ion.android.schedule
 
 import android.os.Bundle
+import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
@@ -10,14 +11,16 @@ import edu.isel.ion.android.R
 /**
  * A simple [Fragment] subclass.
  */
-class CoursesFragment : Fragment() {
+class ScheduleFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_courses, container, false)
+        Log.println(Log.DEBUG,"Schedule Fragment", "Is creating schedule fragment")
+
+        return inflater.inflate(R.layout.fragment_schedule, container, false)
     }
 
 }
