+++ Project/app/src/main/res/layout/classes_list_item.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:padding="20dp"
+    android:padding="@dimen/generic_list_item_padding"
     android:id="@+id/classes_list_item">
 
     <!-- Class name -->
