+++ Project/app/src/main/res/layout/courses_list_item.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:padding="20dp"
+    android:padding="@dimen/generic_list_item_padding"
     android:id="@+id/courses_list_item">
 
     <TextView
