+++ Project/app/src/main/res/layout/curricular_terms_list_item.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:padding="20dp"
+    android:padding="@dimen/generic_list_item_padding"
     android:id="@+id/curricular_terms_list_item">
 
     <TextView
