+++ Project/app/src/main/res/layout/fragment_schedule.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.schedule.ScheduleFragment">
+    tools:context=".schedule.ScheduleFragment">
 
     <TextView
         android:id="@+id/text_schedule"
